[MRP-176] get_ec_devno() in libecho mishandles results
The results from this function are returned by echoing them to stdout. This
needs to be done more safely because other functions that it invokes
(e.g. "remote_shell") can drop detritus into stdout (viz., a terminal control
escape sequence) that mutilates results.

The fundamental bug is that readline(3GNU) should *never* emit terminal
control sequences to stdout.  It should be stderr, especially when stdin is
not a terminal.  We won't be fixing the readline function, but we can mitigate
the bug by dumping unneeded stdout output to stderr or /dev/null.

So, the solution is to clone stdout fd, redirect stdout to an alternate fd
(stderr) and echo results to the cloned fd.

Since this works only as a code fragment with redirected stdout, it must,
perforce, be running in a subshell. The "return" statements need to be "exit"
statements, with the proper success/fail indication. This is not currently
used, but should be more reliable than checking the result count.

Reviewed-by: Alexey Lyashkov <alexey_lyashkov@xyratex.com>