MRP-484 ASSERTION(env->le_ses != NULL) failed
we want to put a CLOSE record to changelog on export destroy for opened files.
it requires to find a full path up to the root. however, objects could be
freed from the cache, so looked up again. the permission check expects the
session parameters to be initialized what is not true for eviction.

ucred cleanup

Reviewed-by: Alexander Zarochentsev <Alexander_Zarochentsev@xyratex.com>
Reviewed-by: Andrew Perepechko <andrew_perepechko@xyratex.com>