CRAY obdfilter: revert xt:29911 and fix autoconf test
Reverts xt:29911 since it is no longer needed.
This mod fixes an autoconf test for
truncate_complete_page. The old test simply grep'd the kernel
source for EXPORT_SYMBOL(truncate_complete_page). Since the
patches are applied to the source with ifdefs this results in a false
positive since the grep doesn't take into account the ifdefs.
This mod replaces the grep test with a try-compile.