LU-1978 llite: Fix i_size race with truncate and mkwrite
LBUG: vvp_io_fault_start() ASSERTION( fio->ft_index <= last )

These changes address a race between the truncate process and the
mkwrite process. We add an additional check in the case of a mkwrite
to make sure we are not past the end of the file. While holding the
i_alloc_sem lock we check the size and make sure we are not past the
end of the file. If we are we return -ENODATA which will trigger a
SIGBUS. This is similar to what occurs in the ext4 filesystem in
these cases. The return status to -ENODATA which will eventually get
mapped to VM_FAULT_NOPAGE (ll_page_mkwrite()). The kernel will then
return a SIGBUS. These changes return -ENODATA to ensure we
correctly clean up. In the case of the mapping being null nothing
will change.

Patch is from LELUS-91 File cray-2.2-LELUS-91.diff, which the 2.2
port of the patch submitted upstream.
Whamcloud bug: LU-1978 http://review.whamcloud.com/4044
 (version for master)

Xyratex-bug-id: MRP-651
Xyratex-bug-id: LELUS-91
Signed-off-by: Wally Wang <wang@cray.com>
Change-Id: I67c46ae93e51b8f3de2b73b8fd156fe73cc8f7cf