LU-107 removal:  Temporarily avoid Lustre startup conflict
Lustre 2.3 has included a failover feature: LU-107 Add scripts for
implementing heartbeat v1 failover.

One of the file, /etc/init.d/lustre, added is in conflict with Cray's
customed legacy Lustre startup script. This problem has been reported
in LU-107 and is pending on resolution. Cray also doesn't use the
failover feature implemented by this patch. So it is safe to
temporarily remove it to avoid conflict and confusion.

Signed-off-by: Wally Wang <wang@cray.com>