LU-1517 ptlrpc: throw net error to ptlrpc for bulk
Client did not propagate errors occurring in the LNet layer while
doing bulk transfers. The specific case observed was an out of memory
condition allocating a tx_phys_buf. The server detected the error and
expected the client to resend the request. But because the client
callback function did not propagate the error, pltrpc_check_set never
resent the request.

The fix is to start reconnect and resend if network error occures for
the bulk transfer.

Signed-off-by: Wally Wang <wang@cray.com>