LU-1565 ptlrpc: resend CANCEL rpc
It is better to deliver CANCEL rpc to server reliably in the following
case: RPC timeout, re-connect, CANCEL resend.  Because server may have
sent BL AST and is waiting for this CANCEL.  This will avoid possible
idle time on server and later client evictions.

CANCEL is always has up-todate lock handle or both enqueue and cancel
are not replayed on recovery, with the exception of the following case:

    BL AST is sent; recovery starts, lock is re-enqueued, BL AST comes
    to client, cancel is created, recovery ends (lock handle has
    changed), CANCEL is sent, its reply gets estale as lock handled
    is not updated in the RPC.

This case is left unfixed and still may result in lock callback timeout
and client eviction, but this race window is much much shorter than
the target case being fixed by this fix.

Also remove lock cancelling from client_common_put_super() as it is
done later in client_disconnect_export().

Xyratex-bug-id: MRP-477
Signed-off-by: Wally Wang <wang@cray.com>