LU-1978 llite: Fix i_size race with truncate and mkwrite
These changes address a race between the truncate process and the
mkwrite process. We add an additional check in the case of a mkwrite
to make sure we are not past the end of the file. While holding the
i_alloc_sem lock we check the size and make sure we are not past the
end of the file. If we are we return -ENODATA which will trigger a
SIGBUS. This is similar to what occurs in the ext4 filesystem in these
cases. The return status to -ENODATA which will eventually get mapped
to VM_FAULT_NOPAGE (ll_page_mkwrite()). The kernel will then return a
SIGBUS. These changes return -ENODATA to ensure we correctly clean
up. In the case of the mapping being null nothing will change.

Xyratex-bug-id: MRP-651
Signed-off-by: Wally Wang <wang@cray.com>