CRAY-0 lnet: Checksum only the bytes transferred.
It is possible for PUTs to have a different length than the
length stored in lntmsg->msg_len since LNET can adjust this
length based on it's buffer size and offset.
lnet_try_match_md() sets the mlength that we use to do the
RDMA transfer.

Therfore we need to compute checksum using
tx->tx_rdma_desc.length and verify the checksum using length
returned in the msg->gnm_u.completion.gncm_retval which contains
the actual number of bytes transmitted.

Signed-off-by: Wally Wang <wang@cray.com>
Change-Id: I04ee272856d1cc081ae14bb285d2767acb0446f4