LU-118 llite: revert to avoid assertion
Revert LU-118 to avoid an assertion it causes. LU-118 is supposed to
be a Lustre workaround for the 767581 kernel BUG. Cray SLES11 SP1
already has a workaround in it to avoid the problem therefore LU-118
is not really needed. It is not clear whether the assertion in LU-118
is "self-inflicted" or a separate problem causing it. Remove LU-118
will probably tell with little risk.

Signed-off-by: Wally Wang <wang@cray.com>
Change-Id: Ia0876dcd0a9c95e47be2fb425b328c62fe69ba4b