CRAY-0 llite: revert xt:29911
This mod fixes an autoconf test for truncate_complete_page. The old
test simply grep'd the kernel source for EXPORT_SYMBOL(
truncate_complete_page). Since the patches are applied to the source
with ifdefs this results in a false positive since the grep doesn't
take into account the ifdefs.  This mod replaces the grep test with a
try-compile.  In addition this mod reverts xt:29911 since it is no
longer needed

Signed-off-by: Wally Wang <wang@cray.com>
Change-Id: If5c0df1db2cfb27620cc9590837479cb60354198