CRAY-0 klnds: Make kgnilnd_bump_timeouts DONE aware
Currently when kgnilnd comes out of quiesce all connections
timeouts are bumped so they dont close from the period they
were paused. kgnilnd_bump_timeouts schedules all the
connections on a peer including ones that are in purgatory in
the GNILND_CONN_DONE state. These connections are not supposed
to be put through the scheduler once they are in the DONE state.

A LBUG can occur if after the quiesce occurs the scheduler
thread does not push the newly scheduled conns through the
state machine fast enough. This can leave DONE conns on the
scheduled list when stack reset is triggered. Stack reset then
puts any scheduled conns through kgnilnd_complete_closed_conn
which when the function sees a conn in the GNILND_CONN_DONE
state it asserts.

Change
1. Add if statement so kgnilnd_bump_timeouts does not
   schedule DONE connections.

Signed-off-by: Wally Wang <wang@cray.com>
Change-Id: I427c1ef78976445724dfe038ca38a4a638ff7c31