CRAY-40958 llite: Change ll_fsync to use *dentry
With 2.x, Lustre added support for different VFS fsync APIs that do
not include a dentry parameter. To make the logic the same in all
cases, the old ll_fsync interface was changed to pull the inode from
the f_dentry in the file parameter. In some cases when using the old
ll_fsync interface, the caller does not set the file parameter
resulting in a NULL dereference. The fix restores the old logic: when
a dentry parameter is provided, get the inode from that parameter
rather than the file parameter.

Signed-off-by: Wally Wang <paf@cray.com>
Change-Id: Id3602e04a77ee76ddb65bf6436fab5e53a9a67f1