CRAY-40969 llite: Fix i_size race with truncate and mkwrite
With the reversion of LU-1030 from 2.3, the wrong version of the patch
for bug 786416 is applied to the 2.3 code. This mod reverts xt:38497
and applies xt:37425. When LU-1030 is applied, the inode is already
locked when vvp_io_fault_start() is called. The back port for Lustre
versions without LU-1030 does the locking within vvp_io_fault_start().

Signed-off-by: Wally Wang <amk@cray.com>
Change-Id: I2f82b280875922475e0342141cee2cdcca657490