CRAY-40174 gnilnd: kgnilnd peer_timeout enhancement
Currently on router nodes kgnilnd peer_health is enabled, when
peer_health is enabled it sets a default timeout factor of
kgn_timeout+kgn_timeout/8. This value currently cannot be adjusted
except through adjust kgn_timeout. This mod allows for the user to
increase the value by setting the module parameter peer_timeout in
conjunction with peer_health.

When peer_timeout is set and peer_health is enabled the timeout passed
to lnet will be what the user has specified as long as it is greater
than the previous fudge calculation. If the user specifies a value
less than fudge kgnilnd will fail to load and throw an error to the
console.

Changes

1. Added module parameter peer_timeout, when peer_health is enabled
   this allows manipulation of the ni_peertimeout value passed to
   lnet.

Signed-off-by: Wally Wang <jshimek@cray.com>
Change-Id: I5b064f8c4da8d03e7ce9e889b073c7538844ed55