Do not panic RHEL 6.7+ client kernels when booting with fips=1
libcfs registers its internal implementations of adler32, CRC32, and
CRC32C as shashes, but this panics RHEL 6.7+ (2.6.32-573.el6.x8 6_64+)
kernels when booting with fips=1. This is due to FIPS not allowing the
use of cryptographic algorithms from non-signed modules. Given that
we can not obtain the RHEL signature key, remove the registration of
the libcfs versions of the algorithms, and use the kernel's
implementation of CRC32C.

This patch addresses the issue only for client builds.

Signed-off-by: Nikitas Angelinas <nikitas.angelinas@seagate.com>