MRP-3249 ptlrpc: do not reduce replay request deadline
New timeout for a replay request sent to a client with early reply is
calculated without taking into account that new deadline for the
replay request on server side is calculated based upon
obd_recovery_timeout. That results in reducing replay deadline which
forces to early disconnect of the client.

Calculate the timeout sent in early reply to a replay request in
according to new deadline of the replay request on server side.

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Change-Id: I40651e4886537cb809c0e1de721632ee180a5554
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/9823
Tested-by: Jenkins
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>
Reviewed-by: Sergey Cheremencev <sergey.cheremencev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>