SNT-443: missing lookup check on the server side
Although, we properly check path traversal, the
final path component is being processed as a
LOOKUP+OPEN intent which gets us into mdt_reint_open()
rather than mdt_reint_getattr(). We do check for
open permission, but we don't check for lookup
permission, which is not safe.

The patch adds a lookup permission check for
the final path component for IT_OPEN requests.

Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>