SNT-446: sda4 Lustre, RHEL 6.9 kernel port
Some ldiskfs and build systems changes to enable a build against
the base RHEL 6.9 kernel, 2.6.32-696.el6.x86_64.

This has been verified to build against the rhel 6.9, in-kernel
IB stack, but fails to build against the Mellanox OFED stack
used in SDA3, so would likely require some additional changes in
order to be useful in Neo; these changes may or may not be
available in the sda3_rhel7.3_client or other branches that
contain IB stack-related modifications.

Signed-off-by: Nikitas Angelinas <nikitas.angelinas@seagate.com>