Proof of concept implementation
Creates a daemon that uses libselinux functionality in order to allow
using a separate policy file for determining AVC decisions for the
Lustre MDS; the daemon and the kernel use a netlink socket for
communications. The current implementation only allows for a single
decision-related netlink message to be in flight at a time, and the
policy to be used in hardcoded, for now.

Signed-off-by: Nikitas Angelinas <nikitas.angelinas@seagate.com>